requirements.txt
torch --index-url https://download.pytorch.org/whl/cu124
torchvision --index-url https://download.pytorch.org/whl/cu124
pandas
matplotlib
numpy
pillow
opencv-python
rich
tqdm
albumentations
wandb